CascadeProjects/image_overlay_tool/requirements.txt
Pillow==10.1.0
tkinter
